[PATCH 1/2] DRM: Add 64-bit support
Some of the data types were incorrectly concatenated
from 64-bit to 32-bit tokens. This patch addresses these issues.

Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-7099
Reviewed-on: https://android.intel.com/160388
Tested-by: Agarwal, Tarun <tarun.agarwal@intel.com>
Reviewed-on: https://android.intel.com:443/291372

+++ shared-core/drm.h
@@ -115,7 +115,7 @@
 #define _DRM_LOCKING_CONTEXT(lock) ((lock) & ~(_DRM_LOCK_HELD|_DRM_LOCK_CONT))
 
 #if defined(__linux__)
-typedef unsigned int drm_handle_t;
+typedef unsigned long int drm_handle_t;
 #else
 #include <sys/types.h>
 typedef unsigned long drm_handle_t;	/**< To mapped regions */
